feat: 新增取得單一貼文api, 修改錯誤 - 在updatedPost, deletedPostyById 新增驗証object id - 在updatedPost, createdPost中新增檢查content是否為空值

+++ controllers/posts.js
@@ -3,6 +3,8 @@ const Post = require("../models/postsModel");
 const handleErrorAsync = require("../service/handleErrorAsync");
 // const handleError = require("../service/handleError");
 const appError = require("../service/appError");
+const validator = require("validator");
+const ObjectId = require("mongoose").Types.ObjectId;
 
 const posts = {
   getPosts: async (req, res, next) => {
@@ -20,10 +22,29 @@ const posts = {
       data: posts,
     });
   },
+  getPostById: async (req, res, next) => {
+    const id = req.params.id;
+    if (!ObjectId.isValid(id)) {
+      return next(appError(400, "查無此id", next));
+    }
+    const post = await Post.findById(id).populate({
+      path: "user",
+      select: "name photo",
+    });
+    if (post !== null) {
+      res.status(200).json({
+        status: "success",
+        data: post,
+      });
+    } else {
+      return next(appError(400, "查無此id", next));
+    }
+  },
   createdPost: handleErrorAsync(async (req, res, next) => {
-    const user = req.user.id;
+    const user = req.user.id; //form isAuth(token)
     const { content, image } = req.body;
-    if (content == undefined) {
+
+    if (!content || validator.isEmpty(content)) {
       return next(appError(400, "你沒有填寫 content 資料", next));
     }
     const newPost = await Post.create({
@@ -38,15 +59,20 @@ const posts = {
   }),
   updatedPost: handleErrorAsync(async (req, res, next) => {
     const id = req.params.id;
-    const data = req.body;
-    if (data.content == undefined) {
+    const { content, image } = req.body;
+
+    if (!ObjectId.isValid(id)) {
+      return next(appError(400, "查無此id", next));
+    }
+
+    if (!content || validator.isEmpty(content)) {
       return next(appError(400, "你沒有填寫 content 資料", next));
     }
     const newPost = await Post.findByIdAndUpdate(
       id,
       {
-        name: data.name,
-        content: data.content,
+        content: content,
+        image: image,
       },
       { new: true }
     );
@@ -61,6 +87,9 @@ const posts = {
   }),
   deletedPostById: handleErrorAsync(async (req, res, next) => {
     const id = req.params.id;
+    if (!ObjectId.isValid(id)) {
+      return next(appError(400, "查無此id", next));
+    }
     const delPost = await Post.findByIdAndDelete(id);
     if (delPost !== null) {
       res.status(200).json({
